DatabaseDBC Class

This class extends the Database class, to query a database using DatabaseConnector with JDBC.

Super classes

DarwinShinyModules::ShinyModule -> DarwinShinyModules::Database -> DatabaseDBC

Methods

Public methods

DatabaseDBC$new()

DatabaseDBC$query()

DatabaseDBC$execute()

DatabaseDBC$uploadTable()

DatabaseDBC$connect()

DatabaseDBC$disconnect()

DatabaseDBC$clone()

Inherited methods

DarwinShinyModules::ShinyModule$UI()

DarwinShinyModules::ShinyModule$getReactiveValues()

DarwinShinyModules::ShinyModule$server()

DarwinShinyModules::ShinyModule$validate()

Method new()

Initializer method

Usage

DatabaseDBC$new(connectionDetails, ...)

Arguments

connectionDetails

(ConnectionDetails) Connection Details from DatabaseConnector::createConnectionDetails

...

Additional parameters to set fields from the ShinyModule parent.

Returns

invisible(self)

Method query()

Query method to query the database using DatabaseConnector::renderTranslateQuerySql

Usage

DatabaseDBC$query(sql, ...)

Arguments

sql

(character(1)) SQL Query

...

Additional parameters for DatabaseConnector::renderTranslateQuerySql

Returns

data.frame

Method execute()

Execute method to execute a query on the database using DatabaseConnector::renderTranslateExecuteSql

Usage

DatabaseDBC$execute(sql, ...)

Arguments

sql

(character(1)) SQL Query to execute

...

Additional parameters for DatabaseConnector::renderTranslateExecuteSql

Returns

invisible(self)

Method uploadTable()

Method to upload data to the database

Usage

DatabaseDBC$uploadTable(tableName, data)

Arguments

tableName

(character(1)) Name of the table

data

(data.frame) data.frame like table to upload

Method connect()

Method to connect to the database.

Usage

DatabaseDBC$connect()

Method disconnect()

Method to disconnect from the database.

Usage

DatabaseDBC$disconnect()

Method clone()

The objects of this class are cloneable with this method.

Usage

DatabaseDBC$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.