DatabaseDBI Class

This class extends the Database class, to query a database using DBI with ODBC.

Super classes

DarwinShinyModules::ShinyModule -> DarwinShinyModules::Database -> DatabaseDBI

Active bindings

tables

(list(n)) List of tbl_dbi tables for the respective driver.

connectArgs

(list(n)) Named list of additional arguments used in DBI::dbConnect()

disconnectArgs

(list(n)) Named list of additional arguments used in DBI::dbDisconnect()

Methods

Public methods

DatabaseDBI$new()

DatabaseDBI$attachTables()

DatabaseDBI$detatchTables()

DatabaseDBI$clone()

Inherited methods

DarwinShinyModules::ShinyModule$UI()

DarwinShinyModules::ShinyModule$server()

DarwinShinyModules::ShinyModule$validate()

Method new()

Initializer method

Usage

DatabaseDBI$new(driver)

Arguments

driver

Driver to use to connect to the database with DBI::dbConnect()

Returns

invisible(self)

Method attachTables()

Method to attach tables from the database in the tables field.

Usage

DatabaseDBI$attachTables(...)

Arguments

...

(character()) Names of tables to attach.

Method detatchTables()

Method to detatch tables from the tables field.

Usage

DatabaseDBI$detatchTables(...)

Arguments

...

(character()) Names of the tables to detatch.

Method clone()

The objects of this class are cloneable with this method.

Usage

DatabaseDBI$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.