EventDuration

Module that displays the Event Duration from the TreatmentPatterns package.

Super class

DarwinShinyModules::ShinyModule -> EventDuration

Active bindings

plot

(PlotPlotly) module.

table

(Table) module.

inputPanel

(InputPanel) module.

summaryEventDuration

(data.table)

cdmSourceInfo

(data.frame)

Methods

Public methods

EventDuration$new()

EventDuration$clone()

Inherited methods

DarwinShinyModules::ShinyModule$UI()

DarwinShinyModules::ShinyModule$server()

DarwinShinyModules::ShinyModule$validate()

Method new()

Initializer method

Usage

EventDuration$new(summaryEventDuration, cdmSourceInfo, ...)

Arguments

summaryEventDuration

(data.frame) summary_event_duration field from the TreatmentPatternsResult object.

cdmSourceInfo

(data.frame) cdm_source_info field from the TreatmentPatternsResult object.

...

Additional parameters to set fields from the ShinyModule parent.

Returns

self

Method clone()

The objects of this class are cloneable with this method.

Usage

EventDuration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

{
  if (interactive()) {
    library(DarwinShinyModules)

    tpr <- TreatmentPatterns::TreatmentPatternsResults$new(
      filePath = "./inst/dummyData/TreatmentPatterns/3.0.0"
    )

    treatmentPathways <- EventDuration$new(
      treatmentPathways = tpr$treatment_pathways,
      dmSourceInfo = tpr$cdm_source_info
    )

    preview(treatmentPathways)
  }
}