Prevalence Module Class

Prevalence module that shows prevalence results from the IncidencePrevalence package.

Value

self

Super class

DarwinShinyModules::ShinyModule -> Prevalence

Active bindings

data

(summarisedResult) SummarisedResult object from Prevalence.

pickers

(list) List of pickers

Methods

Public methods

Prevalence$new()

Prevalence$clone()

Inherited methods

DarwinShinyModules::ShinyModule$UI()

DarwinShinyModules::ShinyModule$getReactiveValues()

DarwinShinyModules::ShinyModule$server()

DarwinShinyModules::ShinyModule$validate()

Method new()

Initializer method

Usage

Prevalence$new(data, ...)

Arguments

data

(summarised_result) Result object from the IncidencePrevalence package.

...

Additional parameters to set fields from the ShinyModule parent.

Method clone()

The objects of this class are cloneable with this method.

Usage

Prevalence$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

{
# \donttest{
 library(DarwinShinyModules)

 if (
   require(
     "IncidencePrevalence",
     character.only = TRUE,
     quietly = TRUE,
     warn.conflicts = FALSE
   )
 ) {
    inc <- omopgenerics::importSummarisedResult(system.file(
      package = "DarwinShinyModules",
      "dummyData/IncidencePrevalence/1.2.0/prevalence.csv"
    ))

    prevMod <- Prevalence$new(data = inc)

    ui <- shiny::fluidPage(
      prevMod$UI()
    )

    server <- function(input, output, session) {
      prevMod$server(input, output, session)
    }

    if (interactive()) {
      shiny::shinyApp(ui = ui, server = server)
    }
  }
# }
}
#> Reading file:
#> /home/runner/work/_temp/Library/DarwinShinyModules/dummyData/IncidencePrevalence/1.2.0/prevalence.csv.
#> Converting to summarised_result:
#> /home/runner/work/_temp/Library/DarwinShinyModules/dummyData/IncidencePrevalence/1.2.0/prevalence.csv.