TreatmentPatterns Module Class

TreatmentPatterns module that shows results from the TreatmentPatterns package.

Super class

ShinyModule -> TreatmentPatterns

Active bindings

analyses

Analyses table from TreatmentPatterns

treatment_pathways

Analyses table from TreatmentPatterns

summary_event_duration

Analyses table from TreatmentPatterns

counts_age

Analyses table from TreatmentPatterns

counts_sex

Analyses table from TreatmentPatterns

counts_year

Analyses table from TreatmentPatterns

attrition

Analyses table from TreatmentPatterns

metadata

Analyses table from TreatmentPatterns

arguments

Analyses table from TreatmentPatterns

cdm_source_info

Analyses table from TreatmentPatterns

Methods

Public methods

TreatmentPatterns$new()

TreatmentPatterns$clone()

Inherited methods

ShinyModule$UI()

ShinyModule$getReactiveValues()

ShinyModule$server()

ShinyModule$validate()

TreatmentPatterns$new()

Initializer method

Usage

TreatmentPatterns$new(...)

Arguments

...

Unnamed TreatmentPatternsResults objects. And Additional parameters to set fields from the ShinyModule parent.

Returns

self

TreatmentPatterns$clone()

The objects of this class are cloneable with this method.

Usage

TreatmentPatterns$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.