PlotPlotly Module Class

Plotly module that handles plotly objects.

Details

Plotly exposes bindings to interact with the plot programaticaly. Currently, only the plotly_selected binding is supported in this module.

Super classes

DarwinShinyModules::ShinyModule -> DarwinShinyModules::Plot -> PlotPlotly

Active bindings

plot

(plotly) object.

source

(character) Source label for the plotly plot.

bindings

(reactivevalues) bindings from the plotly object.

Methods

Public methods

PlotPlotly$clone()

Inherited methods

DarwinShinyModules::ShinyModule$UI()

DarwinShinyModules::ShinyModule$server()

DarwinShinyModules::Plot$initialize()

DarwinShinyModules::Plot$validate()

Method clone()

The objects of this class are cloneable with this method.

Usage

PlotPlotly$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(DarwinShinyModules)

plotlyFun <- function(data) {
  plotly::ggplotly(
    ggplot(data = data, mapping = aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
      geom_point() +
      theme_bw()
  )
}

plotlyModule <- PlotPlotly$new(fun = plotlyFun, args = list(data = iris))

if (interactive()) {
  preview(plotlyModule)
}