TreatmentPatterns

TreatmentPatterns super class. Composed of a PlotWidget and Table modules. This class is an interface and is not meant to be directly used, but to be inherited.

Super class

DarwinShinyModules::ShinyModule -> TreatmentPatterns

Active bindings

data

Data to plot with, usually a data.frame-like object.

widget

(PlotWidget) Module

table

(Table) Module

inputs

(reactiveValues) environment

Methods

Public methods

TreatmentPatterns$new()

TreatmentPatterns$validate()

TreatmentPatterns$UI()

TreatmentPatterns$server()

TreatmentPatterns$clone()

Inherited methods

DarwinShinyModules::ShinyModule$id()

Method new()

Initializer method

Usage

TreatmentPatterns$new(appId, data)

Arguments

appId

(character(1))

data

Data to plot with, usually a data.frame-like object.

Returns

(invisible(self))

Method validate()

Validation method

Usage

TreatmentPatterns$validate()

Returns

(self)

Method UI()

Method to include a tagList to include the body.

Usage

TreatmentPatterns$UI()

Returns

(tagList)

Method server()

Method to handle the back-end.

Usage

TreatmentPatterns$server(input, output, session)

Arguments

input

(input)
Input from the server function.

output

(output)
Output from the server function.

session

(session)
Session from the server function.

Returns

(NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

TreatmentPatterns$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.